# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "app downloads photos from vk.",
        "title": "PhotoDumper",
        "contact": {
            "name": "Rinat Almakhov",
            "url": "https://github.com/Gasoid/"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/",
    "paths": {
        "/albums/{sourceName}/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get albums list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Albums",
                "parameters": [
                    {
                        "type": "string",
                        "description": "source name",
                        "name": "sourceName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "albums",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/download-album/{albumID}/{sourceName}/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "download all photos of particular album",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "download photos of album",
                "parameters": [
                    {
                        "type": "string",
                        "description": "source name",
                        "name": "sourceName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "album ID",
                        "name": "albumID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "directory where photos will be stored",
                        "name": "dir",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/download-all-albums/{sourceName}/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "download all photos of all albums",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "download photos of albums",
                "parameters": [
                    {
                        "type": "string",
                        "description": "source name",
                        "name": "sourceName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "directory where photos will be stored",
                        "name": "dir",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/sources/": {
            "get": {
                "description": "get sources list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Sources",
                "responses": {
                    "200": {
                        "description": "getSources",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "api_key",
            "in": "query"
        }
    }
}